requirements.txt
opencv-python==4.11.0.86
fastapi==0.115.8
Werkzeug==3.1.3
python-multipart==0.0.20
aiofiles==24.1.0
uvicorn==0.34.0
python-dotenv==1.0.1
google-genai==1.3.0
asyncio==3.4.3
loguru==0.7.3
yacs==0.1.8